+++ src/structural_typing/guts/explanations.clj
@@ -23,13 +23,27 @@
 
 ;;; ---
 
-(defn err:only
+(defn err:only-wrong-count
   "The error message produced by `ONLY` when a collection does not have only one element."
-  [collection-with-bad-arity]
-  (cl-format nil "`~S` is supposed to have exactly one element" collection-with-bad-arity))
-(defn oopsie:only [original-path whole-value collection-with-bad-arity]
-  (structural-oopsie original-path whole-value (err:only collection-with-bad-arity)))
-(def as-oopsies:only (pluralize oopsie:only))
+  [original-path collection-with-bad-arity]
+  (cl-format nil "`~S` should be a path through a single-element collection; it passes through `~S`"
+             original-path collection-with-bad-arity))
+(defn oopsie:only-wrong-count [original-path whole-value collection-with-bad-arity]
+  (structural-oopsie original-path whole-value
+                     (err:only-wrong-count original-path collection-with-bad-arity)))
+(def as-oopsies:only-wrong-count (pluralize oopsie:only-wrong-count))
+
+;;; ---
+
+(defn err:some-wrong-count
+  "The error message produced by `SOME` when a collection does not have at least one element"
+  [original-path empty-collection]
+  (cl-format nil "`~S` should be a path to a non-empty collection; it ends in `~S`"
+             original-path empty-collection))
+(defn oopsie:some-wrong-count [original-path whole-value empty-collection]
+  (structural-oopsie original-path whole-value
+                     (err:some-wrong-count original-path empty-collection)))
+(def as-oopsies:some-wrong-count (pluralize oopsie:some-wrong-count))
 
 ;;; ---