+++ test/structural_typing/use/f_ALL_and_missing_values.clj
@@ -54,6 +54,7 @@
 
 (future-fact "same for ONLY")
 (future-fact "and RANGE")
+(future-fact "and SOME")
 
 
 (start-over!)