+++ CHANGELOG.md
@@ -7,6 +7,7 @@ See [here](http://keepachangelog.com/) for the change log format.
   map to traverse to build up paths.
 - ADD: Strings can now be used in paths. Like keywords, they denote map keywords to follow.
 - ADD: `ONLY` is like `ALL`, but it insists that its collection have exactly one element.
+- ADD: `SOME` is like `ALL`, but it insists that its collection have at least one element.
 - ADD: By default, a non-function `x` used in predicate position behaves as `(exactly x)`.
        - Regular expressions are treated as in Midje.
        - Comparison of record to map as in Midje.